refactor + test fix (password test)
