+++ src/main/java/com/parachute/booking/admin/AdminMapper.java
@@ -1,24 +1,28 @@
 package com.parachute.booking.admin;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
 @Component
+@RequiredArgsConstructor
 public class AdminMapper {
 
-    public AdminDto mapAdminDto(Admin admin){
+    private final PasswordEncoder passwordEncoder;
+
+    public AdminDto mapAdminDto(Admin admin) {
         return AdminDto.builder()
                 .email(admin.getEmail())
                 .login(admin.getLogin())
-                .password(admin.getPassword())
                 .id(admin.getId())
                 .build();
     }
 
-    public Admin mapAdmin(AdminDto adminDto){
+    public Admin mapAdmin(AdminDto adminDto) {
         return Admin.builder()
                 .email(adminDto.getEmail())
                 .login(adminDto.getLogin())
-                .password(adminDto.getPassword())
+                .password(passwordEncoder.encode(adminDto.getPassword()))
                 .build();
     }
 }
