+++ src/main/java/com/parachute/booking/security/CounterController.java
@@ -1,38 +1,55 @@
 package com.parachute.booking.security;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.web.bind.annotation.GetMapping;
 
-@EnableWebSecurity
-@EnableGlobalMethodSecurity(prePostEnabled = true)
-@RequiredArgsConstructor
-public class CounterController extends WebSecurityConfigurerAdapter {
+import javax.servlet.http.Cookie;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+import java.util.Arrays;
+import java.util.Map;
 
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
+public class CounterController {
+
+    @GetMapping("/counter")
+    Map<Object, Object> getCounter(HttpServletRequest request, HttpServletResponse response) {
+
+        int counter = 0;
+
+        Cookie[] cookies = request.getCookies();
+        if (cookies == null) {
+            counter = 1;
+            Cookie cookie = new Cookie("counter", String.valueOf(counter));
+            response.addCookie(cookie);
+            return Map.of("counter", counter);
+        }
+
+        Cookie cookie = Arrays.stream(cookies)
+                .filter(c -> c.getName().equals("counter"))
+                .findFirst()
+                .orElseGet(() -> new Cookie("counter", "0"));
+
+        String stringvalue = cookie.getValue();
+        counter = Integer.parseInt(stringvalue);
+        counter++;
+        cookie.setValue(String.valueOf(counter));
+        cookie.setMaxAge(1800);
+        response.addCookie(cookie);
+        return Map.of("counter", counter);
     }
 
-    @Override
-    protected void configure(HttpSecurity httpSecurity) throws Exception {
-        httpSecurity
-                .authorizeRequests()
-                .anyRequest().permitAll()
-                .and()
-                .formLogin()
-                .and()
-                .logout()
-                .and()
-                .headers().frameOptions().disable()
-                .and()
-                .csrf().disable()
-                .httpBasic();
-    }
+    @GetMapping("/counter/session")
+    Map<Object, Object> getSessionCounter(HttpSession httpSession) {
+        Object counter = httpSession.getAttribute("counter");
 
+        if (counter == null) {
+            httpSession.setAttribute("counter", 1);
+            return Map.of("counter", 1);
+        }
+        int counterInt = (int) counter;
+        counterInt++;
+        httpSession.setAttribute("counter", counterInt);
+
+        return Map.of("counter", counterInt);
+    }
 }
